Added some restrictions for free version of plugin

+++ includes/class-pattern-builder-controller.php
@@ -2,6 +2,7 @@
 
 require_once __DIR__ . '/class-pattern-builder-abstract-pattern.php';
 
+global $pb_fs;
 class Pattern_Builder_Controller
 {
 	public function get_pb_block_post_for_pattern($pattern)
@@ -57,67 +58,72 @@ class Pattern_Builder_Controller
 
 	public function update_theme_pattern(Abstract_Pattern $pattern)
 	{
-		// get the pb_block post if it already exists
-		$post = get_page_by_path(sanitize_title($pattern->name), OBJECT, 'pb_block');
+		if (pb_fs()->can_use_premium_code__premium_only()) {
 
-		if (empty($post)) {
-			// if it doesn't exist, check if a wp_block post exists
-			// this is for any user patterns that are being converted to theme patterns
-			// It will be converted to a pb_block post when it is updated
-			$post = get_page_by_path(sanitize_title($pattern->name), OBJECT, 'wp_block');
+			// get the pb_block post if it already exists
+			$post = get_page_by_path(sanitize_title($pattern->name), OBJECT, 'pb_block');
+
+			if (empty($post)) {
+				// if it doesn't exist, check if a wp_block post exists
+				// this is for any user patterns that are being converted to theme patterns
+				// It will be converted to a pb_block post when it is updated
+				$post = get_page_by_path(sanitize_title($pattern->name), OBJECT, 'wp_block');
+			}
+
+			if (empty($post)){
+				// create a new post if it doesn't exist
+				$post = $this->create_pb_block_post_for_pattern($pattern);
+			}
+
+
+			wp_update_post([
+				'ID'           => $post->ID,
+				'post_title'   => $pattern->title,
+				'post_content' => $pattern->content,
+				'post_excerpt' => $pattern->description,
+				'post_type'    => 'pb_block',
+			]);
+
+			if ($pattern->synced) {
+				delete_post_meta($post->ID, 'wp_pattern_sync_status');
+			} else {
+				update_post_meta($post->ID, 'wp_pattern_sync_status', 'unsynced');
+			}
+
+			if ($pattern->keywords) {
+				update_post_meta($post->ID, 'wp_pattern_keywords', implode(',', $pattern->keywords));
+			} else {
+				delete_post_meta($post->ID, 'wp_pattern_keywords');
+			}
+
+			if ($pattern->blockTypes) {
+				update_post_meta($post->ID, 'wp_pattern_block_types', implode(',', $pattern->blockTypes));
+			} else {
+				delete_post_meta($post->ID, 'wp_pattern_block_types');
+			}
+
+			if ($pattern->templateTypes) {
+				update_post_meta($post->ID, 'wp_pattern_template_types', implode(',', $pattern->templateTypes));
+			} else {
+				delete_post_meta($post->ID, 'wp_pattern_template_types');
+			}
+
+			if ($pattern->postTypes) {
+				update_post_meta($post->ID, 'wp_pattern_post_types', implode(',', $pattern->postTypes));
+			} else {
+				delete_post_meta($post->ID, 'wp_pattern_post_types');
+			}
+
+			// store categories
+			wp_set_object_terms($post->ID, $pattern->categories, 'wp_pattern_category', false);
+
+			// update the pattern file
+			$this->update_theme_pattern_file($pattern);
+
+			return $pattern;
 		}
 
-		if (empty($post)){
-			// create a new post if it doesn't exist
-			$post = $this->create_pb_block_post_for_pattern($pattern);
-		}
-
-
-		wp_update_post([
-			'ID'           => $post->ID,
-			'post_title'   => $pattern->title,
-			'post_content' => $pattern->content,
-			'post_excerpt' => $pattern->description,
-			'post_type'    => 'pb_block',
-		]);
-
-		if ($pattern->synced) {
-			delete_post_meta($post->ID, 'wp_pattern_sync_status');
-		} else {
-			update_post_meta($post->ID, 'wp_pattern_sync_status', 'unsynced');
-		}
-
-		if ($pattern->keywords) {
-			update_post_meta($post->ID, 'wp_pattern_keywords', implode(',', $pattern->keywords));
-		} else {
-			delete_post_meta($post->ID, 'wp_pattern_keywords');
-		}
-
-		if ($pattern->blockTypes) {
-			update_post_meta($post->ID, 'wp_pattern_block_types', implode(',', $pattern->blockTypes));
-		} else {
-			delete_post_meta($post->ID, 'wp_pattern_block_types');
-		}
-
-		if ($pattern->templateTypes) {
-			update_post_meta($post->ID, 'wp_pattern_template_types', implode(',', $pattern->templateTypes));
-		} else {
-			delete_post_meta($post->ID, 'wp_pattern_template_types');
-		}
-
-		if ($pattern->postTypes) {
-			update_post_meta($post->ID, 'wp_pattern_post_types', implode(',', $pattern->postTypes));
-		} else {
-			delete_post_meta($post->ID, 'wp_pattern_post_types');
-		}
-
-		// store categories
-		wp_set_object_terms($post->ID, $pattern->categories, 'wp_pattern_category', false);
-
-		// update the pattern file
-		$this->update_theme_pattern_file($pattern);
-
-		return $pattern;
+		return new WP_Error('premium_required', 'Saving Theme Patterns requires the premium version of Pattern Builder.', ['status' => 403]);
 	}
 
 	/**
@@ -133,10 +139,14 @@ class Pattern_Builder_Controller
 
 		if (empty($post)) {
 			// check if the pattern exists in the database as a pb_block post
-			// this is for any user patterns that are being converted to theme patterns
+			// this is for any user patterns that are being converted from theme patterns
 			// It will be converted to a wp_block post when it is updated
 			$post = get_page_by_path($pattern->name, OBJECT, 'pb_block');
 			$convert_from_theme_pattern = true;
+		}
+
+		if ( $convert_from_theme_pattern && !pb_fs()->can_use_premium_code() ) {
+			return new WP_Error('premium_required', 'Converting Theme Patterns to User Patterns requires the premium version of Pattern Builder.', ['status' => 403]);
 		}
 
 		if (empty($post)) {
@@ -245,26 +255,31 @@ class Pattern_Builder_Controller
 
 	public function delete_theme_pattern(Abstract_Pattern $pattern)
 	{
-		$path = $this->get_pattern_filepath($pattern);
+		if (pb_fs()->can_use_premium_code__premium_only()) {
 
-		if (!$path) {
-			return new WP_Error('pattern_not_found', 'Pattern not found', ['status' => 404]);
+			$path = $this->get_pattern_filepath($pattern);
+
+			if (!$path) {
+				return new WP_Error('pattern_not_found', 'Pattern not found', ['status' => 404]);
+			}
+
+			$deleted = unlink($path);
+
+			if (!$deleted) {
+				return new WP_Error('pattern_delete_failed', 'Failed to delete pattern', ['status' => 500]);
+			}
+
+			$pb_block_post = $this->get_pb_block_post_for_pattern($pattern);
+			$deleted = wp_delete_post($pb_block_post->ID, true);
+
+			if (!$deleted) {
+				return new WP_Error('pattern_delete_failed', 'Failed to delete pattern', ['status' => 500]);
+			}
+
+			return ['message' => 'Pattern deleted successfully'];
 		}
 
-		$deleted = unlink($path);
-
-		if (!$deleted) {
-			return new WP_Error('pattern_delete_failed', 'Failed to delete pattern', ['status' => 500]);
-		}
-
-		$pb_block_post = $this->get_pb_block_post_for_pattern($pattern);
-		$deleted = wp_delete_post($pb_block_post->ID, true);
-
-		if (!$deleted) {
-			return new WP_Error('pattern_delete_failed', 'Failed to delete pattern', ['status' => 500]);
-		}
-
-		return ['message' => 'Pattern deleted successfully'];
+		return new WP_Error('premium_required', 'Deleting Theme Patterns requires the premium version of Pattern Builder.', ['status' => 403]);
 	}
 
 	public function update_theme_pattern_file(Abstract_Pattern $pattern)
